Устранена утечка памяти при пересоздании volatileImage главного меню. Оптимизирована библиотека FoxGUI, исправлено отсутствие иконки оповещения.

+++ src/main/java/game/freya/entities/Hero.java
@@ -68,7 +68,7 @@ public class Hero {
 
     @Builder.Default
     @Column(name = "speed")
-    private short speed = 6;
+    private byte speed = 6;
 
     @Builder.Default
     @Enumerated(EnumType.STRING)
